requirements.txt
cellpose
numpy
opencv-python
streamlit
openai
pillow
matplotlib
scikit-image
